+++ packages/api/brevo-api/eslint.config.mjs
@@ -1,16 +1,12 @@
+import { defineConfig, globalIgnores } from "eslint/config";
 import eslintConfigNestJs from "@comet/eslint-config/nestjs.js";
 
-/** @type {import('eslint')} */
-const config = [
-    {
-        ignores: ["src/db/migrations/**", "dist/**", "src/**/*.generated.ts", "lib/**", "block-meta.json"],
-    },
+export default defineConfig([
+    globalIgnores(["src/db/migrations/**", "dist/**", "src/**/*.generated.ts", "lib/**", "block-meta.json"]),
     ...eslintConfigNestJs,
     {
         rules: {
             "@comet/no-other-module-relative-import": "off",
         },
     },
-];
-
-export default config;
+]);