+++ packages/eslint-config/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/eslint-config
 
-<<<<<<< HEAD
 ## 9.0.0-beta.0
 
 ### Major Changes
@@ -21,7 +20,7 @@
 
 - Updated dependencies [db6b83a]
     - @comet/eslint-plugin@9.0.0-beta.0
-=======
+
 ## 8.20.0
 
 ### Patch Changes
@@ -39,7 +38,6 @@
 ### Patch Changes
 
 - @comet/eslint-plugin@8.18.0
->>>>>>> main
 
 ## 8.17.1